feat: rule engine and smt
